update Android compile SDK and build configurations
This commit updates the Android `compileSdk` to version 36 and `targetSdk` to 35. It also increments the `versionName` to "1.0.1".

Additionally, it introduces signing configurations for release builds, enabling different keystore setups for CI/CD environments (GitHub Actions) and local development. Proguard rules and minification are now enabled for release builds.

The `buildFeatures.compose` flag is set to true and `composeOptions.kotlinCompilerExtensionVersion` is specified as "1.5.10".